services/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow>=11.0.0
opencv-python==4.13.0.92
numpy==2.4.2
sqlalchemy==2.0.46
psycopg2-binary==2.9.9
bcrypt==4.1.1
PyJWT==2.8.1
alembic==1.13.0
pytesseract==0.3.13
